Move some functions into Settings and PlayerRecords instead of main
