+++ src/main.rs
@@ -1,16 +1,16 @@
 use player_records::PlayerRecords;
-use std::io::ErrorKind;
-use std::path::{Path, PathBuf};
+
+use std::path::Path;
 use std::time::Duration;
 use steamapi::steam_api_loop;
 use steamid_ng::SteamID;
 use tokio::sync::mpsc::UnboundedSender;
 
 use clap::{ArgAction, Parser};
-use demo::demo_loop;
+// use demo::demo_loop;
 use io::{Commands, IOManager};
 use launchoptions::LaunchOptions;
-use settings::{ConfigFilesError, Settings};
+use settings::Settings;
 use state::SharedState;
 use tappet::SteamAPI;
 use tracing_appender::non_blocking::WorkerGuard;
@@ -18,7 +18,7 @@ use tracing_subscriber::{
     fmt::writer::MakeWriterExt, layer::SubscriberExt, util::SubscriberInitExt, EnvFilter, Layer,
 };
 
-mod demo;
+// mod demo;
 mod gamefinder;
 mod io;
 mod launchoptions;
@@ -28,7 +28,7 @@ mod server;
 mod settings;
 mod state;
 mod steamapi;
-mod web;
+// mod web;
 
 #[derive(Parser, Debug)]
 #[command(author, version, about, long_about = None)]
@@ -72,53 +72,8 @@ fn main() {
     let args = Args::parse();
 
     // Load settings
-    let settings_path: PathBuf = args
-        .config
-        .as_ref()
-        .map(|i| Ok(i.into()))
-        .unwrap_or(Settings::locate_config_file_path()).map_err(|e| {
-            tracing::error!("Could not find a suitable location for the configuration: {}\nPlease specify a file path manually with --config", e);
-        }).unwrap_or(PathBuf::from("config.yaml"));
-
-    let mut settings = match Settings::load_from(settings_path, &args) {
-        Ok(settings) => settings,
-        Err(ConfigFilesError::Yaml(path, e)) => {
-            tracing::error!("{} could not be loaded: {:?}", path, e);
-            tracing::error!(
-                "Please resolve any issues or remove the file, otherwise data may be lost."
-            );
-            panic!("Failed to load configuration")
-        }
-        Err(ConfigFilesError::IO(path, e)) if e.kind() == ErrorKind::NotFound => {
-            tracing::warn!("Could not locate {}, creating new configuration.", &path);
-            let mut settings = Settings::default();
-            settings.set_config_path(path.into());
-            settings.set_overrides(&args);
-            settings
-        }
-        Err(e) => {
-            tracing::error!("Could not load configuration: {:?}", e);
-            tracing::error!(
-                "Please resolve any issues or remove the file, otherwise data may be lost."
-            );
-            panic!("Failed to load configuration")
-        }
-    };
-
+    let settings = Settings::load_or_create(&args);
     settings.save_ok();
-
-    // Locate TF2 directory
-    match gamefinder::locate_tf2_folder() {
-        Ok(tf2_directory) => {
-            settings.set_tf2_directory(tf2_directory);
-        }
-        Err(e) => {
-            if args.tf2_dir.is_none() {
-                tracing::error!("Could not locate TF2 directory: {:?}", e);
-                tracing::error!("If you have a valid TF2 installation you can specify it manually by appending ' --tf2_dir \"Path to Team Fortress 2 folder\"' when running the program.");
-            }
-        }
-    }
 
     // Launch options and overrides
     let launch_opts = match LaunchOptions::new(
@@ -182,39 +137,7 @@ fn main() {
     // Just some settings we'll need later
     let port = settings.get_port();
 
-    // Playerlist
-    let playerlist_path: PathBuf = args
-        .playerlist
-        .as_ref()
-        .map(|i| Ok(i.into()))
-        .unwrap_or(PlayerRecords::locate_playerlist_file()).map_err(|e| {
-            tracing::error!("Could not find a suitable location for the playerlist: {} \nPlease specify a file path manually with --playerlist otherwise information may not be saved.", e); 
-        }).unwrap_or(PathBuf::from("playerlist.json"));
-
-    let playerlist = match PlayerRecords::load_from(playerlist_path) {
-        Ok(playerlist) => playerlist,
-        Err(ConfigFilesError::Json(path, e)) => {
-            tracing::error!("{} could not be loaded: {:?}", path, e);
-            tracing::error!(
-                "Please resolve any issues or remove the file, otherwise data may be lost."
-            );
-            panic!("Failed to load playerlist")
-        }
-        Err(ConfigFilesError::IO(path, e)) if e.kind() == ErrorKind::NotFound => {
-            tracing::warn!("Could not locate {}, creating new playerlist.", &path);
-            let mut playerlist = PlayerRecords::default();
-            playerlist.set_path(path.into());
-            playerlist
-        }
-        Err(e) => {
-            tracing::error!("Could not load playerlist: {:?}", e);
-            tracing::error!(
-                "Please resolve any issues or remove the file, otherwise data may be lost."
-            );
-            panic!("Failed to load playerlist")
-        }
-    };
-
+    let playerlist = PlayerRecords::load_or_create(&args);
     playerlist.save_ok();
 
     // Get vars from settings before it is borrowed
@@ -240,12 +163,14 @@ fn main() {
             }
 
             // Spawn web server
+            /*
             {
                 let state = state.clone();
                 tokio::spawn(async move {
                     web::web_main(state, port).await;
                 });
             }
+            */
 
             if autolaunch_ui {
                 if let Err(e) = open::that(Path::new(&format!("http://localhost:{}", port))) {
@@ -263,6 +188,7 @@ fn main() {
             }
 
             // Demo manager
+            /*
             {
                 if args.demo_monitoring {
                     let demo_path = state.settings.read().get_tf2_directory().join("tf/demos");
@@ -273,6 +199,7 @@ fn main() {
                     });
                 }
             }
+            */
 
             // Main and refresh loop
             {
